In stop cmd, replace region regex with actual list of avail. regions
stop.go had a regex to extract the region from the node name. This list was missing some
regions (il-* e.g.) and could also match on non-valid regions, if we had a node tagged
tailout which such. Instead, use the list of regions as returned by DescribeRegions to
search for matches in the node name.